change field type, fix all bugs known

+++ src/main/java/Common/Entity/BasketItem.java
@@ -18,28 +18,28 @@ import lombok.Setter;
 @Getter
 public class BasketItem {
 
-    @JsonProperty("SellerId") private long sellerId;
-    @JsonProperty("ProductId") private long productId;
+    @JsonProperty("SellerId") private int sellerId;
+    @JsonProperty("ProductId") private int productId;
 
     @JsonProperty("ProductName") private String productName;
-    @JsonProperty("UnitPrice") private double unitPrice;
-    @JsonProperty("FreightValue") private double freightValue; //运费
+    @JsonProperty("UnitPrice") private float unitPrice;
+    @JsonProperty("FreightValue") private float freightValue; //运费
 
     @JsonProperty("Quantity") private int quantity;
-    @JsonProperty("Voucher") private double vouchers;
+    @JsonProperty("Voucher") private float vouchers;
     @JsonProperty("Version") private int version;
 
     public BasketItem() {
     }
 
     @JsonCreator
-    public BasketItem(@JsonProperty("SellerId") long sellerId,
-                      @JsonProperty("ProductId") long productId,
+    public BasketItem(@JsonProperty("SellerId") int sellerId,
+                      @JsonProperty("ProductId") int productId,
                       @JsonProperty("ProductName") String productName,
-                      @JsonProperty("UnitPrice") double unitPrice,
-                      @JsonProperty("FreightValue") double freightValue,
+                      @JsonProperty("UnitPrice") float unitPrice,
+                      @JsonProperty("FreightValue") float freightValue,
                       @JsonProperty("Quantity") int quantity,
-                      @JsonProperty("Voucher") double vouchers,
+                      @JsonProperty("Voucher") float vouchers,
                       @JsonProperty("Version") int version
                       ) {
         this.sellerId = sellerId;
